[FE] refactor: type 분리 (#31)
