+++ src/frontend/src/components/common/Modal/index.tsx
@@ -1,20 +1,9 @@
 import React from 'react';
 
 import useModalStore from '../../../stores/modalStore';
+import { ModalType } from '../../../types';
 
 import * as S from './styles';
-
-type ValueOf<T> = T[keyof T];
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const ModalTypes = {
-  BASIC: 'basic',
-  WITH_FOOTER: 'withFooter',
-};
-export type ModalType = ValueOf<typeof ModalTypes>;
-export interface BaseModalData {
-  content: React.ReactNode;
-}
 
 interface ModalProps {
   children: React.ReactNode;
