+++ src/frontend/src/components/guild/types/index.ts
@@ -1,0 +1,1 @@
+export type CreateGuildStep = 'initial' | 'customize';
